bukutamu/tamu/utils.py
import requests
from django.conf import settings

def send_telegram_notification(message):
    """
    Mengirim notifikasi ke Telegram saat ada pesan baru
    """
    bot_token = settings.TELEGRAM_BOT_TOKEN
    chat_id = settings.TELEGRAM_CHAT_ID
    
    url = f"https://api.telegram.org/bot{bot_token}/sendMessage"
    data = {
        "chat_id": chat_id,
        "text": message,
        "parse_mode": "HTML"
    }
    
    try:
        response = requests.post(url, data=data)
        return response.json()
    except Exception as e:
        print(f"Error sending Telegram notification: {e}")
        return None 